App: implement draft of an acceleration based spin
This commit replaces the "advance by one" spin logic with a "physics" based
approach.
We model the wheel as an object with position, speed and acceleration
and apply the acceleration to the speed using a completely arbitary
formula that probably needs to be replaced with something better.

The new logic is float based, so we floor it to the lowest integer
before advancing our list by that many elements.

The end result is painfully slow in big lists, but it's an ok enough
start.